refactor: extract completed button logic to custom hook
