+++ src/pages/DetailPage/components/CompletedButton/CompletedButton.tsx
@@ -1,54 +1,20 @@
-import { FC, useEffect, useState } from 'react';
+import { FC } from 'react';
 import { styled } from 'styled-components';
 import { Theme } from '../../../../styles';
 import { ClearIcon, HelpIcon } from '../../../../assets';
-import { useDetailData } from '../../container';
-import { useNavigate } from 'react-router-dom';
-import { toggleDoneData } from '../../../../api';
+import { useCompleted, useDetailData } from '../../container';
 type CompletedButtonProps = {};
 
 export const CompletedButton: FC<CompletedButtonProps> = ({}) => {
-  const { data, isLoading, error } = useDetailData();
-  const [localDone, setLocalDone] = useState<boolean | null>(null);
-  const [localDoneCount, setLocalDoneCount] = useState<number | null>(null);
-  const navigate = useNavigate();
+  const { isLoading, error, done, doneCount } = useDetailData();
+  const { handleClickDoneButton, localDoneCount } = useCompleted(
+    done,
+    doneCount,
+  );
 
-  if (isLoading) {
+  if (isLoading || error) {
     return <></>;
   }
-
-  const { done, doneCount, postId } = data;
-
-  const handleClickDoneButton = async () => {
-    if (localDoneCount === 5 && !localDone) {
-      alert('이미 해결된 민원글입니다.');
-      navigate('/main');
-      return;
-    }
-
-    // console.log('loginstate', isLoggedInState);
-    // TODO
-    const token = localStorage.getItem('token');
-
-    if (token) {
-      await toggleDoneData(postId);
-
-      if (localDone) {
-        setLocalDoneCount((prevCount) => (prevCount ? prevCount - 1 : 0));
-        setLocalDone(false);
-      } else {
-        setLocalDoneCount((prevCount) => (prevCount ? prevCount + 1 : 1));
-        setLocalDone(true);
-      }
-    } else {
-      alert('로그인 후 이용 가능합니다');
-    }
-  };
-
-  useEffect(() => {
-    setLocalDone(done);
-    setLocalDoneCount(doneCount);
-  }, [done, doneCount]);
 
   return (
     <StWrapper>
@@ -142,125 +108,3 @@ export const StButtonTextWrapper = styled.div`
   font-size: ${Theme.fontSizes.h2};
   font-weight: ${Theme.fontWeights.h2};
 `;
-
-{
-  /* <div
-  style={{
-    width: 280,
-    height: 51,
-    paddingTop: 0.25,
-    paddingBottom: 0.75,
-    boxShadow: '0px 0px 5px rgba(41, 47, 61, 0.25)',
-    justifyContent: 'center',
-    alignItems: 'center',
-    display: 'inline-flex',
-  }}
->
-  <div style={{ width: 280, position: 'relative' }}>
-    <div
-      style={{
-        width: 202,
-        height: 50,
-        paddingLeft: 12,
-        paddingRight: 12,
-        paddingTop: 6,
-        paddingBottom: 6,
-        left: 0,
-        top: 0,
-        position: 'absolute',
-        background: '#5782FA',
-        borderTopLeftRadius: 10,
-        borderTopRightRadius: 10,
-        justifyContent: 'center',
-        alignItems: 'center',
-        gap: 10,
-        display: 'inline-flex',
-      }}
-    >
-      <div style={{ width: 178, height: 32, position: 'relative' }}>
-        <div
-          style={{
-            width: 144,
-            height: 32,
-            padding: 6,
-            left: 0,
-            top: 0,
-            position: 'absolute',
-            justifyContent: 'center',
-            alignItems: 'center',
-            display: 'inline-flex',
-          }}
-        >
-          <div
-            style={{
-              flex: '1 1 0',
-              alignSelf: 'stretch',
-              textAlign: 'center',
-              color: '#FEFEFE',
-              fontSize: 18,
-              fontFamily: 'Pretendard',
-              fontWeight: '700',
-              wordWrap: 'break-word',
-            }}
-          >
-            해결된 민원이에요
-          </div>
-        </div>
-        <div
-          style={{
-            width: 32,
-            height: 32,
-            left: 146,
-            top: 0,
-            position: 'absolute',
-            flexDirection: 'column',
-            justifyContent: 'flex-start',
-            alignItems: 'flex-start',
-            display: 'inline-flex',
-          }}
-        >
-          <img
-            style={{ width: 28, height: 28 }}
-            src="https://via.placeholder.com/28x28"
-          />
-        </div>
-      </div>
-    </div>
-    <div
-      style={{
-        width: 78,
-        height: 50,
-        paddingLeft: 12,
-        paddingRight: 12,
-        paddingTop: 6,
-        paddingBottom: 6,
-        left: 202,
-        top: 0,
-        position: 'absolute',
-        opacity: 0.6,
-        background: '#FEFEFE',
-        borderTopLeftRadius: 100,
-        borderTopRightRadius: 100,
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        gap: 10,
-        display: 'inline-flex',
-      }}
-    >
-      <div
-        style={{
-          flex: '1 1 0',
-          textAlign: 'center',
-          color: '#14171F',
-          fontSize: 18,
-          fontFamily: 'Pretendard',
-          fontWeight: '700',
-          wordWrap: 'break-word',
-        }}
-      >
-        2/5
-      </div>
-    </div>
-  </div>
-</div>; */
-}
