+++ src/pages/DetailPage/container/hooks/index.ts
@@ -1,2 +1,4 @@
 export * from './detailData';
 export * from './modifyData';
+export * from './metoo';
+export * from './completed';
